+++ ClipboardMonitorLite.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipboardMonitorLite", "ClipboardMonitorLite\ClipboardMonitorLite.csproj", "{F7AA0F1C-CB0C-4DF3-A473-34B0D5EEF3E6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SettingsLib", "SettingsLib\SettingsLib.csproj", "{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SettingsLib", "SettingsLib\SettingsLib.csproj", "{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudMessageLib", "CloudMessageLib\CloudMessageLib.csproj", "{54171BE8-D12E-4406-9753-613AEF55385B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{16CDBE54-7D1C-4F57-8ABF-7E8520B95B49}.Release|Any CPU.Build.0 = Release|Any CPU
+		{54171BE8-D12E-4406-9753-613AEF55385B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54171BE8-D12E-4406-9753-613AEF55385B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54171BE8-D12E-4406-9753-613AEF55385B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54171BE8-D12E-4406-9753-613AEF55385B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
