+++ ClipboardMonitorLite/ClipboardMonitorLite.csproj
@@ -182,47 +182,16 @@
   <ItemGroup>
     <Compile Include="ClipboardActions\AutoHistoryCleaner.cs" />
     <Compile Include="Cloud\CheckConnection.cs" />
-    <Compile Include="Cloud\InboundMessage.cs" />
     <Compile Include="Cloud\CloudInteractions.cs" />
     <Compile Include="Cloud\OnlineState.cs" />
-    <Compile Include="Cloud\OutgoingMessage.cs" />
     <Compile Include="Exceptions\ExceptionHandling.cs" />
     <Compile Include="FileOperations\LaunchOnStartup.cs" />
     <Compile Include="FormControls\FormControl.cs" />
     <Compile Include="FormControls\FormEvents.cs" />
     <Compile Include="Languages\LanguageControl\LanguageCode.cs" />
+    <Compile Include="Languages\LanguageControl\LanguageOnForm.cs" />
     <Compile Include="Languages\lang_da.Designer.cs">
       <DependentUpon>lang_da.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_sk.Designer.cs">
-      <DependentUpon>lang_sk.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_nl.Designer.cs">
-      <DependentUpon>lang_nl.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_fil.Designer.cs">
-      <DependentUpon>lang_fil.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_sr.Designer.cs">
-      <DependentUpon>lang_sr.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_fi.Designer.cs">
-      <DependentUpon>lang_fi.resx</DependentUpon>
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-    </Compile>
-    <Compile Include="Languages\lang_no.Designer.cs">
-      <DependentUpon>lang_no.resx</DependentUpon>
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
     </Compile>
@@ -231,12 +200,46 @@
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
     </Compile>
-    <Compile Include="Languages\lang_hu.Designer.cs">
+    <Compile Include="Languages\lang_fi.Designer.cs">
+      <DependentUpon>lang_fi.resx</DependentUpon>
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
-      <DependentUpon>lang_hu.resx</DependentUpon>
     </Compile>
-    <Compile Include="Languages\LanguageControl\LanguageOnForm.cs" />
+    <Compile Include="Languages\lang_fil.Designer.cs">
+      <DependentUpon>lang_fil.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_hu.Designer.cs">
+      <DependentUpon>lang_hu.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_nl.Designer.cs">
+      <DependentUpon>lang_nl.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_no.Designer.cs">
+      <DependentUpon>lang_no.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_pl.Designer.cs">
+      <DependentUpon>lang_pl.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_sk.Designer.cs">
+      <DependentUpon>lang_sk.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
+    <Compile Include="Languages\lang_sr.Designer.cs">
+      <DependentUpon>lang_sr.resx</DependentUpon>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+    </Compile>
     <Compile Include="Main_Form.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -263,38 +266,41 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>lang_da.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_sk.resx">
+    <EmbeddedResource Include="Languages\lang_en.resx">
       <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_sk.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_nl.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_nl.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_fil.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_fil.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_pl.resx" />
-    <EmbeddedResource Include="Languages\lang_sr.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_sr.Designer.cs</LastGenOutput>
+      <LastGenOutput>lang_en.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Include="Languages\lang_fi.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>lang_fi.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_no.resx">
+    <EmbeddedResource Include="Languages\lang_fil.resx">
       <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_no.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Languages\lang_en.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>lang_en.Designer.cs</LastGenOutput>
+      <LastGenOutput>lang_fil.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Include="Languages\lang_hu.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>lang_hu.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Languages\lang_nl.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>lang_nl.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Languages\lang_no.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>lang_no.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Languages\lang_pl.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>lang_pl.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Languages\lang_sk.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>lang_sk.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Languages\lang_sr.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>lang_sr.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Include="Main_Form.resx">
       <DependentUpon>Main_Form.cs</DependentUpon>
@@ -365,6 +371,10 @@
   </ItemGroup>
   <ItemGroup />
   <ItemGroup>
+    <ProjectReference Include="..\CloudMessageLib\CloudMessageLib.csproj">
+      <Project>{54171be8-d12e-4406-9753-613aef55385b}</Project>
+      <Name>CloudMessageLib</Name>
+    </ProjectReference>
     <ProjectReference Include="..\SettingsLib\SettingsLib.csproj">
       <Project>{16cdbe54-7d1c-4f57-8abf-7e8520b95b49}</Project>
       <Name>SettingsLib</Name>
