+++ CloudMessageLib/OutgoingMessage.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel;
 
-namespace ClipboardMonitorLite.Cloud
+namespace CloudMessageLib
 {
-    public class InboundMessage : INotifyPropertyChanged
+    public class OutgoingMessage : INotifyPropertyChanged
     {
         public event PropertyChangedEventHandler PropertyChanged;
 
@@ -33,6 +33,8 @@ namespace ClipboardMonitorLite.Cloud
                 InvokePropertyChanged(new PropertyChangedEventArgs("MachineName"));
             }
         }
+
+
 
         private void InvokePropertyChanged(PropertyChangedEventArgs e)
         {
